# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Category:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  dto.CreateProductDTO:
    properties:
      categories:
        items:
          $ref: '#/definitions/domain.Category'
        type: array
      description:
        type: string
      name:
        type: string
      price:
        type: number
    required:
    - name
    - price
    type: object
  dto.CreateReviewDTOAdmin:
    properties:
      productID:
        type: string
      rating:
        type: integer
      text:
        type: string
      userID:
        type: string
    type: object
  dto.CreateReviewDTOUser:
    properties:
      rating:
        type: integer
      text:
        type: string
    type: object
...
  dto.UpdateProductDTO:
    properties:
      categories:
        items:
          $ref: '#/definitions/domain.Category'
        type: array
      description:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
...
  v1.failure:
    properties:
      error:
        $ref: '#/definitions/v1.failureInfo'
    type: object
  v1.failureInfo:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: invalid request body
        type: string
    type: object
  v1.success:
    properties:
      data: {}
    type: object
host: localhost:8080
info:
  contact:
    email: paw1a@yandex.ru
    name: API Support
    url: https://t.me/paw1a
  description: This is simple api of e-commerce shop
  title: E-commerce API
  version: "1.0"
paths:
...
  /admins/products:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.success'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Get all products
      tags:
      - admin-products
    post:
      consumes:
      - application/json
      parameters:
      - description: product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Create product
      tags:
      - admin-products
  /admins/products/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: product id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Delete product
      tags:
      - admin-products
    get:
      consumes:
      - application/json
      parameters:
      - description: product id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Get product by id
      tags:
      - admin-products
    put:
      consumes:
      - application/json
      parameters:
      - description: product id
        in: path
        name: id
        required: true
        type: string
      - description: product update fields
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProductDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Update product
      tags:
      - admin-products
  /admins/products/{id}/reviews:
    get:
      consumes:
      - application/json
      parameters:
      - description: product id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Get product reviews list
      tags:
      - admin-products
  /admins/reviews:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.success'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Get all reviews
      tags:
      - admin-reviews
    post:
      consumes:
      - application/json
      parameters:
      - description: review
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/dto.CreateReviewDTOAdmin'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Create review
      tags:
      - admin-reviews
  /admins/reviews/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: review id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Delete review
      tags:
      - admin-reviews
    get:
      consumes:
      - application/json
      parameters:
      - description: review id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Get review by id
      tags:
      - admin-reviews
...
  /products:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.success'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      summary: Get all products
      tags:
      - products
  /products/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: product id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      summary: Get product by id
      tags:
      - products
  /products/{id}/reviews:
    get:
      consumes:
      - application/json
      parameters:
      - description: product id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      summary: Get product reviews list
      tags:
      - products
    post:
      consumes:
      - application/json
      parameters:
      - description: product id
        in: path
        name: id
        required: true
        type: string
      - description: review
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/dto.CreateReviewDTOUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - UserAuth: []
      summary: Create review
      tags:
      - products
...
securityDefinitions:
  AdminAuth:
    in: header
    name: Authorization
    type: apiKey
  UserAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"